requirements.txt
google-generativeai
gtts
pydub
librosa
soundfile
moviepy==1.0.3
google-auth-oauthlib